Expose a handler for deleting stored messages.
To avoid looking up the home node in Dynamo, only messages that have
been written to storage are deleted. Direct updates are not removed.
This seems fine, since deleting a message sent to a connected client
is inherently racy: the message might be delivered before the
connection node receives the cancellation.