Bug 1215253 - Part 2: Make DELETE endpoints return 404 when uaid/chid not found. r=jconlin
The important part of this patch is registering the channel with the
message router (via register_channel).

Before this, we produced an endpoint, but did not keep track of the chid
at all.  Adding the chid exposes an issue where DynamoDB deletes fail
because the `DELETE chids :chid` clause tries to remove an unknown
chid.  (This might only when running a local moto server.)  In any case,
this patch tests that the chid is known before we try to delete.

I would have preferred to unconditionally delete, ignore errors, and
return 404 based on the delete having succeeded -- but I couldn't figure
out how to achieve this.  So, extra look-up it is!