feat: track last month connected using last_connect router field
Utilizes the existing router last_connect global index to track the first
connect of the month for the client. The last_connect number is a number
joining the year/month/hour with a random int to reduce hot key contention
on the index.

Additional integration tests were added to verify the last_connect is updated
appropriately for simplepush and webpush clients.

This changes the get_rotating_message_table function to automatically create
the table if needed, reducing boiler-plate in the settings object wrapping
that behavior.

Registration endpoint handlers saving router data all update the last_connect
as well (new channels do not update the value).

Closes #253