Pass router data as args
Noted in the PR, I was holding the credentials in the router object.
There is a chance that the credentials might be changed if two
devices were registering at the same time. It's safer to pass the
router data as an argument in the callback series, but it does
require touching a lot more code.