bug: overwrite records during monthly migration to avoid disconnect edge-case
During monthly rotation, if a user drops the connection after the monthly
rotation has begun, but before it completed, then the message table will be
updated but not the router table. This means that the next call will fail
since put_item wasn't called with overwrite=True. This patch fixes that
and includes an integration test to verify the fix.

Closes #320