fix: allow webpush w/no ttl & cleanup 400 logging
Allows webpush notifications with no TTL header included as long
as the client is online. Otherwise a 400 is returned with a
descriptive URL including a link to the appropriate docs on how to
include a proper TTL.

Refactors how the TTL setup code is handled, and cleans up the 400
logging for consistency.

Closes #357, #358