fix: Correct documents to strongly recommend well formatted UUIDs
The code normalizes UUIDs used for channelIDs and UIADs to lower case,
dashed format. The docs now state that and the code will reject UUIDs
that are not well formatted. This is important for clients that may
normalize differently.

BREAKING CHANGE: This patch will reject UUIDs that are not expressed as
lower case, dash formatted (e.g. "00001111-aaaa-bbbb-cccc-222233334444")

Closes #392