feat: refactor simplepush endpoint for validation schemas
This refactor splits out the base handler and several bits of
generic callbacks into a base.py module. Each web handler will
now get its own module, with simplepush.py as the initial
refactor candidate.

Marshmallow schemas are used to validate incoming requests and
format/coerce/validate data in a separate thread. This allows
the validation to make AWS calls or other expensive calls that
previously required punting calls back/forth between the twisted
thread-pool.

The SimplePush endpoints issued are now under a new spush prefix
so that they can go to a separate handler entirely that has the
schema validation. WebPush endpoints will similarly be sent to
a new prefix. An additional toggle is stored on the user object
to track whether their record was upgraded for this change so
that at a later point we can expire-on-connect old user records.

Issue #379