fix: handle different behavior in rotating table tests at month-end
Two of the tests for update_rotating_tables failed to account for
a different code-path that runs at the end of the month when the
next day is a new month. This change populates the next month
message table entry so that it won't matter if its the end of
the month or not for the tests to run properly.

Closes #502