fix: refactor register uaid lookup to verify proper records
When a client says hello, the register user method now performs a router lookup
to verify the record actually exists and has the appropriate necessary attributes.
A check is also done for webpush users to verify the server still has a table object
it can use to pull notifications for during verification so the uaid can be reset if
needed.

This cleans up the hello handling, removing the need for a separate hello finish
and simplifying  the callbacks.

Closes #523